+++ src/scene/mod.rs
@@ -88,7 +88,7 @@ impl NavMeshContainer {
     }
 
     /// Creates a handle to navmesh from its index.
-    pub fn handle_from_index(&self, i: usize) -> Handle<Navmesh> {
+    pub fn handle_from_index(&self, i: u32) -> Handle<Navmesh> {
         self.pool.handle_from_index(i)
     }
