+++ src/scene2d/graph.rs
@@ -131,13 +131,13 @@ impl Graph {
             .set_position(Vector2::default());
     }
 
-    pub fn capacity(&self) -> usize {
+    pub fn capacity(&self) -> u32 {
         self.pool.get_capacity()
     }
 
     /// Makes new handle from given index. Handle will be none if index was either out-of-bounds
     /// or point to a vacant pool entry.
-    pub fn handle_from_index(&self, index: usize) -> Handle<Node> {
+    pub fn handle_from_index(&self, index: u32) -> Handle<Node> {
         self.pool.handle_from_index(index)
     }
